N3: extract the run report so the CLI can emit it without FastAPI
The HTML run report could only be produced by the dashboard endpoint, so CI
could not get the artifact without standing up the web app. The payload
builder, renderer and filename move to livekit_diag/run_report.py, which
imports no web framework and touches no storage, and voicegw livekit report
exports a stored run through it. diagnostics.py drops from 1535 to 455
lines and keeps its old private names as bindings, so existing callers and
tests still work.

One renderer, proven two ways. The endpoint coroutine is called directly
and its bytes compared against the file the CLI wrote for the same run,
with the clock frozen on run_report only: if either side still had its own
copy, only one would be frozen and generated_at would differ. Backed by an
identity assertion that diagnostics._render_report_html is
run_report.render_html.

Self-containment is scanned from the bytes on disk rather than the
response, and the list is stricter than the endpoint's own test: no script,
link, img, iframe, object, embed, svg, @import, url(, src, srcset,
integrity, crossorigin, and no http:// or https:// at all.

schema_version is 1 before and after. A failed run exports at exit 0 and
reads as measured-nothing: no numeric cells, no reply-latency rows, and the
JSON twin carries nulls rather than zeros. Exiting non-zero on a FAIL run
would discard the report that explains it, so gating stays check's job.
